changed to ulong

+++ tango/io/vfs/FileFolder.d
@@ -369,7 +369,7 @@ class FileGroup : VfsFiles
 
         ***********************************************************************/
 
-        final size_t bytes ()
+        final ulong bytes ()
         {
                 return stats.bytes;
         }
